# docs/reviews/project-category-only-evidence/current-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List owned Projects",
        "description": "createdAt DESC, id DESC. Query is a case-insensitive literal substring of name or stack. Cursor binds all filters and limit; no cross-page snapshot guarantee.",
        "operationId": "list_1",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "enum": [
                "all",
                "unity",
                "server"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ]
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered cursor page with total matching count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR or INVALID_CURSOR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED, or CSRF_INVALID for an authenticated mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create an owned Project",
        "description": "Requires session CSRF/Origin checks. Retry the same Idempotency-Key and body within 24 hours to replay the original 201 response. Category omission preserves the legacy v1 fingerprint; supplied null/UUID uses v2, including raw UUID case. JSON property order is ignored; supplied values and field presence are preserved. New responses include nullable categoryId; pre-deployment replay retains the original shape without categoryId. Replay never resolves current Category state.",
        "operationId": "create_1",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Session token from GET /api/v1/auth/csrf",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "[!-~]+"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created, or original creation response replayed",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProjectResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyProjectResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED, or CSRF_INVALID for an authenticated mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible Category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED or REVISION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get an owned Project, including archived Projects",
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owned Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED, or CSRF_INVALID for an authenticated mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND (missing or inaccessible)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update, archive or unarchive an owned Project",
        "description": "Only supplied fields change. categoryId omission preserves, null clears, UUID assigns an owned Category independently of scope; other explicit nulls are invalid. Revision is required and advances once per successful PATCH, including same-value updates. Empty strings clear optional text. Archived Projects remain editable; archive/unarchive preserves Category and connected data.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Session token from GET /api/v1/auth/csrf",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Full updated Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED, or CSRF_INVALID for an authenticated mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND (missing or inaccessible)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "REVISION_CONFLICT; reload and reconcile the draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "fieldErrors",
          "message",
          "requestId"
        ]
      },
...
      "CreateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Unknown fields are rejected. Explicit null is allowed only for categoryId. Omitted/null categoryId creates uncategorized; a UUID assigns an owned Category independently of scope. Omitted optional strings default to empty; progress defaults to zero. String limits use UTF-16 code units after applicable trimming.",
        "example": {
          "name": "Example Project",
          "scope": "server",
          "stack": "Java",
          "progress": 12.5
        },
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "subtitle": {
            "type": "string",
            "default": "",
            "maxLength": 4000,
            "minLength": 0
          },
          "scope": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ],
            "pattern": "unity|server"
          },
          "stack": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "progress": {
            "type": "number",
            "default": 0,
            "description": "Manual progress; decimals allowed",
            "maximum": 100,
            "minimum": 0
          },
          "currentMilestone": {
            "type": "string",
            "default": "",
            "description": "Goal memo, independent of milestone resources",
            "maxLength": 200,
            "minLength": 0
          },
          "repositoryUrl": {
            "type": "string",
            "default": "",
            "description": "Empty or absolute http/https URL with a host",
            "maxLength": 2000
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "scope",
          "stack"
        ]
      },
      "ProjectResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "subtitle": {
            "type": "string",
            "maxLength": 4000
          },
          "scope": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ]
          },
          "stack": {
            "type": "string",
            "maxLength": 200
          },
          "progress": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "currentMilestone": {
            "type": "string",
            "maxLength": 200
          },
          "repositoryUrl": {
            "type": "string",
            "maxLength": 2000
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "colorToken": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ],
            "readOnly": true
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "categoryId",
          "colorToken",
          "createdAt",
          "currentMilestone",
          "id",
          "name",
          "progress",
          "repositoryUrl",
          "revision",
          "scope",
          "stack",
          "status",
          "subtitle",
          "updatedAt"
        ]
      },
      "LegacyProjectResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "subtitle": {
            "type": "string",
            "maxLength": 4000
          },
          "scope": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ]
          },
          "stack": {
            "type": "string",
            "maxLength": 200
          },
          "progress": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "currentMilestone": {
            "type": "string",
            "maxLength": 200
          },
          "repositoryUrl": {
            "type": "string",
            "maxLength": 2000
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "colorToken": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ],
            "readOnly": true
          }
        },
        "required": [
          "colorToken",
          "createdAt",
          "currentMilestone",
          "id",
          "name",
          "progress",
          "repositoryUrl",
          "revision",
          "scope",
          "stack",
          "status",
          "subtitle",
          "updatedAt"
        ]
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Only supplied fields change. categoryId omission preserves, null clears, UUID assigns an owned Category; archived Projects are editable. Other explicit nulls and unknown fields are rejected. Revision-only and same-value updates increment revision once.",
        "example": {
          "revision": 1,
          "status": "archived"
        },
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          },
          "name": {
            "type": "string",
            "description": "Nonblank after trim when supplied",
            "maxLength": 100,
            "minLength": 1
          },
          "subtitle": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "scope": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ],
            "pattern": "unity|server"
          },
          "stack": {
            "type": "string",
            "description": "Nonblank after trim when supplied",
            "maxLength": 200,
            "minLength": 1
          },
          "progress": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "currentMilestone": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "repositoryUrl": {
            "type": "string",
            "maxLength": 2000
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ],
            "pattern": "active|archived"
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "revision"
        ]
      },
...
      "ProjectListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "All matching owned rows, not the page size"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Null on the last page"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "total"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}